+++ routes/notes.js
@@ -11,20 +11,25 @@ const router = express.Router();
 
 router.get('/', (req, res, next) => {
 
-  const { searchTerm, folderId } = req.query;
+  const { searchTerm, folderId, tagId } = req.query;
   let filter = {};
 
   if (searchTerm) {
     const regex = new RegExp(searchTerm, 'i');
-    filter.$or =[{ 'title': regex}, {'content': regex}];
+    filter.$or = [{ 'title': regex }, { 'content': regex }];
   }
 
   if (folderId) {
     filter.folderId = folderId;
   }
 
+  if (tagId) {
+    filter.tags = tagId;
+  }
+
   Note.find(filter)
-    .sort({updatedAt: 'desc'})
+    .populate('tags')
+    .sort({ updatedAt: 'desc' })
     .then(results => {
       if (results) {
         res.json(results);
@@ -49,6 +54,7 @@ router.get('/:id', (req, res, next) => {
   }
 
   Note.findById(noteId)
+    .populate('tags')
     .then(results => {
       if (results) {
         res.json(results);
@@ -64,24 +70,45 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
 
-  const { title, content } = req.body;
+  const { title, content, folderId, tags = [] } = req.body;
 
   const newItem = {
     title: title,
-    content: content
+    content: content,
+    folderId: folderId,
+    tags: tags
   };
 
-  if(!newItem.title) {
+  if (newItem.folderId === '') {
+    delete newItem.folderId;
+  }
+
+  if (!newItem.title) {
     const err = new Error('Missing `title` in request body');
     err.status = 400;
     return next(err);
   }
 
+  if (folderId && !mongoose.Types.ObjectId.isValid(folderId)) {
+    const err = new Error('The folder id is not valid');
+    err.status = 400;
+    return next(err);
+  }
+
+  tags.forEach(tagId => {
+    if (tagId && !mongoose.Types.ObjectId.isValid(tagId)) {
+      const err = new Error('There is an invalid tag id');
+      err.status = 400;
+      return next(err);
+    }
+  });
+
   Note.create(newItem)
-    .then(results => {
-      res.location(`${req.originalUrl}/${results.id}`)
+    .then(note => note.populate('tags').execPopulate())
+    .then(populatedNote => {
+      res.location(`${req.originalUrl}/${populatedNote.id}`)
         .status(201)
-        .json(results);
+        .json(populatedNote);
     })
     .catch(err => {
       if (err.code === 11000) {
@@ -96,22 +123,45 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
 
   const updateId = req.params.id;
-  const updateObj = {};
-  const updatableFields = ['title', 'content'];
+  const { title, content, folderId, tags = [] } = req.body;
 
-  updatableFields.forEach(field => {
-    if (field in req.body) {
-      updateObj[field] = req.body[field];
+  if (!mongoose.Types.ObjectId.isValid(updateId)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
+
+
+  if (folderId && !mongoose.Types.ObjectId.isValid(folderId)) {
+    const err = new Error('The `folderId` is not valid');
+    err.status = 400;
+    return next(err);
+  }
+
+  if (tags.length) {
+    const badTags = tags.filter((tag) => !mongoose.Types.ObjectId.isValid(tag));
+    if (badTags.length) {
+      const err = new Error('There is an invalid `tagId`');
+      err.status = 400;
+      return next(err);
     }
-  });
+  }
 
-  if(!updateObj.title) {
+  if (!title) {
     const err = new Error('Missing `title` in request body');
     err.status = 400;
     return next(err);
   }
 
-  Note.findByIdAndUpdate(updateId, {$set: updateObj}, {new: true})
+  const updateObj = { title, content, folderId, tags };
+
+  if (updateObj.folderId === '') {
+    delete updateObj.folderId;
+    updateObj.$unset = { folderId: '' };
+  }
+
+  Note.findByIdAndUpdate(updateId, { $set: updateObj }, { new: true })
+    .populate('tags')
     .then(results => {
       res.json(results);
     })
@@ -128,6 +178,12 @@ router.put('/:id', (req, res, next) => {
 router.delete('/:id', (req, res, next) => {
 
   const deleteId = req.params.id;
+
+  if (!mongoose.Types.ObjectId.isValid(deleteId)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
 
   Note.findByIdAndDelete(deleteId)
     .then(res.sendStatus(204))
